💬 Add PDF/DOCX format tips to upload busy messages
